feat: further split down the area layer codebase
